commsgaze: Removed TextToSpeechInit and made TTS initialisation as a part of Initialisation
